+++ packages/classes/classes.ts
@@ -1,0 +1,3 @@
+export * from './src/messenger';
+export * from './src/sdk-service';
+export * from './src/config';