+++ packages/content/content.ts
@@ -1,4 +1,3 @@
-export * from './src/api-endpoints';
 export * from './src/content-store-service';
 export * from './src/navigation-service';
 export * from './src/url-transformation-service';