+++ packages/utils/src/types.ts
@@ -1,0 +1,5 @@
+import { PartialObserver } from 'rxjs/index';
+
+export declare type ObserverOrNext<T> = (value: T) => void | PartialObserver<T>;
+
+export declare type SearchEngines = 'solr' | 'elastic';