+++ packages/utils/utils.ts
@@ -1,2 +1,3 @@
 export * from './src/misc';
 export * from './src/logging';
+export * from './src/types';